make mybool use ints on the teensy, makes more sense on a 32bit mcu

+++ mybool.h
@@ -2,7 +2,7 @@
 #define MYBOOL_H
 #include <stdint.h>
 
-typedef uint8_t bool;
+typedef int bool;
 #define true 1
 #define false 0
